+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Bump core-ktx to 1.8.0
 - Fix `RetentionType` getting obfuscated by R8/Proguard
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
+- [In Progress: 03 Jun 2022] Implement `OverdueAppointmentListItemNew` adapter
 
 ### Changes